fix: unicite des epreuves + repartition points float + admin peut promouvoir un prof
- upload-exam (file + json) : verification d'unicite avec used_combinations
- Rejet avec erreur claire si pas assez de combinaisons uniques
- Points en float (NUMERIC) pour accepter les decimales (6.67, etc.)
- Redistribution automatique des points garantissant somme = total_score
- Nouvel endpoint PUT /admin/teachers/{id}/role
- Nouvelle page AdminTeachers avec boutons Promouvoir/Restrograder
- Migration SQL : migration_points_float.sql

+++ server/core/db.py
@@ -299,7 +299,7 @@ def get_session_exercises(session_id: int) -> list[dict]:
     return result.data or []
 
 
-def add_session_exercise(session_id: int, exercise_id: int, sort_order: Optional[int] = None, points_override: Optional[int] = None) -> Optional[dict]:
+def add_session_exercise(session_id: int, exercise_id: int, sort_order: Optional[int] = None, points_override: Optional[float] = None) -> Optional[dict]:
     """Ajoute un exercice a une session. Retourne None si deja present."""
     supabase = get_supabase()
     data = {
